requirements.txt
term-image==0.7.2 
google-genai==1.12.1
python-dotenv==1.1.0
